[Kernel] move module tolibc/libdl.
