+++ components/libc/libdl/dlerror.c
@@ -1,15 +1,11 @@
 /*
- * File      : dlerror.c
- * This file is part of RT-Thread RTOS
- * COPYRIGHT (C) 2006 - 2010, RT-Thread Development Team
+ * Copyright (c) 2006-2018, RT-Thread Development Team
  *
- * The license and distribution terms for this file may be
- * found in the file LICENSE in this distribution or at
- * http://www.rt-thread.org/license/LICENSE
- *
+ * SPDX-License-Identifier: Apache-2.0
+ * 
  * Change Logs:
- * Date           Author		Notes
- * 2010-11-17      yi.qiu	first version
+ * Date           Author      Notes
+ * 2010-11-17     yi.qiu      first version
  */
 
 #include <rtthread.h>
@@ -19,6 +15,4 @@ const char *dlerror(void)
 {
 	return "TODO";
 }
-
 RTM_EXPORT(dlerror)
-
