+++ components/libc/libdl/dlfcn.h
@@ -1,3 +1,13 @@
+/*
+ * Copyright (c) 2006-2018, RT-Thread Development Team
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * 
+ * Change Logs:
+ * Date           Author      Notes
+ * 2010-11-17     yi.qiu      first version
+ */
+
 #ifndef __DLFCN_H_
 #define __DLFCN_H_
 
@@ -16,4 +26,3 @@ void *dlsym(void *handle, const char *symbol);
 int dlclose (void *handle);
 
 #endif
-
