+++ components/libc/libdl/dlopen.c
@@ -1,11 +1,7 @@
 /*
- * File      : dlopen.c
- * This file is part of RT-Thread RTOS
- * COPYRIGHT (C) 2006 - 2010, RT-Thread Development Team
+ * Copyright (c) 2006-2018, RT-Thread Development Team
  *
- * The license and distribution terms for this file may be
- * found in the file LICENSE in this distribution or at
- * http://www.rt-thread.org/license/LICENSE
+ * SPDX-License-Identifier: Apache-2.0
  *
  * Change Logs:
  * Date           Author        Notes
@@ -16,11 +12,13 @@
 #include <rtm.h>
 #include <string.h>
 
+#include "dlmodule.h"
+
 #define MODULE_ROOT_DIR     "/modules"
 
 void* dlopen(const char *filename, int flags)
 {
-    rt_module_t module;
+    struct rt_dlmodule *module;
     char *fullpath;
     const char*def_path = MODULE_ROOT_DIR;
 
@@ -40,11 +38,21 @@ void* dlopen(const char *filename, int flags)
         fullpath = (char*)filename; /* absolute path, use it directly */
     }
 
-    /* find in module list */
-    module = rt_module_find(fullpath);
+    rt_enter_critical();
 
-    if(module != RT_NULL) module->nref++;
-    else module = rt_module_open(fullpath);
+    /* find in module list */
+    module = dlmodule_find(fullpath);
+
+    if(module != RT_NULL) 
+    {
+        rt_exit_critical();
+        module->nref++;
+    }
+    else 
+    {
+        rt_exit_critical();
+        module = dlmodule_load(fullpath);
+    }
 
     if(fullpath != filename)
     {
@@ -54,4 +62,3 @@ void* dlopen(const char *filename, int flags)
     return (void*)module;
 }
 RTM_EXPORT(dlopen);
-
